♻️ Reorganize tag feature layout
