+++ packages/server/src/features/tag/graphql/index.ts
@@ -1,0 +1,5 @@
+export { tagFieldResolvers } from './tag.field.resolver.js';
+export { createTagMutationResolver, tagMutationResolvers } from './tag.mutation.resolver.js';
+export { tagQueryResolvers } from './tag.query.resolver.js';
+export { tagResolvers } from './tag.resolver.js';
+export { tagTypeDefs } from './tag.type-defs.js';
