+++ packages/server/src/views/index.ts
@@ -1,4 +1,3 @@
 export * from './auth.js';
 export * from './image.js';
 export * from './mcp-admin.js';
-export * from './tag.js';
